Abstract common query
